Clean up clippy warnings (cargo clippy now silent)
collapsible_if (connection.rs let-chain), derivable Default on
TransactionBuilder, is_multiple_of in xdr align, and an unnecessary i64 cast in
the new civil-date code.

+++ src/value.rs
@@ -86,7 +86,7 @@ pub struct CivilTimestamp {
 fn civil_from_unix_days(z: i64) -> CivilDate {
     let z = z + 719_468;
     let era = (if z >= 0 { z } else { z - 146_096 }) / 146_097;
-    let doe = (z - era * 146_097) as i64; // dia da era [0, 146096]
+    let doe = z - era * 146_097; // dia da era [0, 146096]
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365; // [0, 399]
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
